+++ src/checkbox-group/checkbox-group.ts
@@ -11,19 +11,27 @@ export default class CheckBoxGroup extends SuperComponent {
   relations = {
     '../checkbox/checkbox': {
       type: 'descendant' as 'descendant',
-      linked() {
-        this.updateChildren();
+      linked(child) {
+        const { value, disabled } = this.data;
+
+        child.setData({
+          checked: value.includes(child.data.value),
+          disabled: disabled || child.data.disabled,
+        });
       },
     },
   };
 
   data = {
+    prefix,
     classPrefix: name,
     checkboxOptions: [],
+    indeterminate: false,
   };
 
   properties = {
     ...Props,
+    customStyle: String,
     defaultValue: {
       type: null,
       value: undefined,
@@ -31,14 +39,14 @@ export default class CheckBoxGroup extends SuperComponent {
   };
 
   observers = {
-    value: function () {
+    value() {
       this.updateChildren();
     },
   };
 
   lifetimes = {
     attached() {
-      this.handleCreateMulCheckbox();
+      this.initWithOptions();
     },
   };
 
@@ -53,127 +61,103 @@ export default class CheckBoxGroup extends SuperComponent {
     getChilds() {
       let items = this.getRelationNodes('../checkbox/checkbox');
       if (!items.length) {
-        items = this.selectAllComponents('.t-checkbox-option');
+        items = this.selectAllComponents(`.${prefix}-checkbox-option`);
       }
       return items || [];
     },
 
-    // slot插入选项
     updateChildren() {
       const items = this.getChilds();
+      const { value, options } = this.data;
 
-      const { value, disabled } = this.data;
       if (items.length > 0) {
         items.forEach((item: any) => {
           !item.data.checkAll &&
             item.setData({
-              checked: value?.indexOf(item.data.value) > -1,
+              checked: value?.includes(item.data.value),
             });
-          item.setDisabled(disabled);
         });
         // 关联可全选项
-        if (items.findIndex((item) => item.data.checkAll) > -1) {
-          items.forEach((item) => {
-            item.setOptionLinked(true);
-          });
-          this.handleHalfCheck(items.length);
+        if (options.some((item) => item.checkAll)) {
+          this.handleCheckall();
         }
       }
     },
-    updateValue({ name, checked }) {
+
+    updateValue({ key, checked }) {
       const { value, max } = this.data;
       let newValue = value;
-      if (max && checked && newValue.length === max) {
-        return;
-      }
+      const keySet = new Set(this.getChilds().map((item) => item.data.value));
+
+      newValue = newValue.filter((value) => keySet.has(value));
+
+      if (max && checked && newValue.length === max) return;
+
       if (checked) {
-        newValue = newValue.concat(name);
+        newValue = newValue.concat(key);
       } else {
-        const index = newValue.findIndex((v: string) => v === name);
+        const index = newValue.findIndex((v: string) => v === key);
         newValue.splice(index, 1);
       }
       this._trigger('change', { value: newValue });
     },
-    // 支持自定义options
-    handleCreateMulCheckbox() {
-      const { options } = this.data;
-      // 数字数组｜字符串数组｜对像数组
-      if (!options?.length || !Array.isArray(options)) {
-        return;
-      }
-      const optionsValue = [];
-      try {
-        options.forEach((element) => {
-          const typeName = typeof element;
-          if (typeName === 'number' || typeName === 'string') {
-            optionsValue.push({
-              label: `${element}`,
-              value: element,
-            });
-          } else if (typeName === 'object') {
-            optionsValue.push({
-              ...element,
-            });
-          }
-        });
-        this.setData({
-          checkboxOptions: optionsValue,
-        });
-        this.updateChildren();
-      } catch (error) {
-        console.error('error', error);
-      }
-    },
-    // 处理全选
-    handleCheckAll(e) {
-      const { checked, option, name } = e.detail || e;
-      const items = this.getChilds();
 
-      if (!option) {
-        if (!items?.length) {
-          return;
-        }
-        this._trigger('change', {
-          value: items
-            .map((item) => {
-              if (item.data.disabled) {
-                return this.data.value.includes(item.data.value) ? item.data.value : '';
-              }
-              return checked && !item.data.checkAll ? item.data.value : '';
-            })
-            .filter((val) => val),
-        });
+    initWithOptions() {
+      const { options } = this.data;
+
+      if (!options?.length || !Array.isArray(options)) return;
+
+      const checkboxOptions = options.map((item) => {
+        const isLabel = ['number', 'string'].includes(typeof item);
+        return isLabel
+          ? {
+              label: `${item}`,
+              value: item,
+            }
+          : { ...item };
+      });
+
+      this.setData({
+        checkboxOptions,
+      });
+    },
+
+    handleInnerChildChange(e) {
+      const { item } = e.target.dataset;
+      const { checked } = e.detail;
+      const { checkboxOptions, indeterminate } = this.data;
+
+      if (item.checkAll) {
+        const value =
+          !checked && indeterminate
+            ? checkboxOptions.map((item) => item.value)
+            : checkboxOptions
+                .filter((item) => {
+                  if (item.disabled) {
+                    return this.data.value.includes(item.value);
+                  }
+                  return checked && !item.checkAll;
+                })
+                .map((item) => item.value);
+        this._trigger('change', { value });
       } else {
-        this.updateValue({ name, checked });
+        this.updateValue({ key: item.value, checked });
       }
     },
-    // 处理options半选
-    handleHalfCheck(len: number) {
-      const items = this.getChilds();
-      const checkboxOptions = items.filter((i) => !i.data.checkAll);
-      const all = checkboxOptions.map((item) => item.data.value);
-      const enableValue = checkboxOptions
-        .filter((i) => !i.data.disabled)
-        .map((item) => item.data.value);
-      const currentVal = Array.from(new Set(this.data.value?.filter((i) => all.indexOf(i) > -1)));
-      const element = items.find((item) => item.data.checkAll);
-      if (currentVal.length) {
-        element?.setData({ checked: true });
-        element?.changeCheckAllHalfStatus(currentVal.length !== len - 1);
-        // 取消全选
-        element?.setCancel(enableValue.every((val) => currentVal.includes(val)));
-      } else {
-        element?.setData({ checked: false });
-      }
-    },
-    // 设置可全选option选项
-    handleOptionLinked() {
-      const items = this.selectAllComponents('.t-checkbox-option');
-      if (this.data.checkboxOptions.length) {
-        items.forEach((item) => {
-          item.setOptionLinked(true);
-        });
-      }
+
+    handleCheckall() {
+      const { checkboxOptions, value } = this.data;
+      const valueSet = new Set(value);
+      const isCheckall = checkboxOptions.every((item) => (item.checkAll ? true : valueSet.has(item.value)));
+      const items = this.selectAllComponents(`.${prefix}-checkbox-option`);
+      const $target = items.find((item) => item.data.checkAll);
+
+      $target.setData({
+        checked: valueSet.size > 0,
+      });
+      this.setData({
+        indeterminate: !isCheckall,
+      });
     },
   };
 }