+++ src/dropdown-menu/dropdown-menu.ts
@@ -1,76 +1,90 @@
-import TComponent from '../common/component';
+import { RelationsOptions, SuperComponent, wxComponent } from '../common/src/index';
 import config from '../common/config';
+import props from './props';
+import type { TdDropdownMenuProps } from './type';
 
 const { prefix } = config;
 const name = `${prefix}-dropdown-menu`;
 
-TComponent({
-  properties: {
-    overlay: {
-      type: Boolean,
-      value: true,
-    },
-    duration: {
-      type: Number,
-      value: 200,
-    },
-    closeOnClickOverlay: {
-      type: Boolean,
-      value: true,
-    },
-  },
-  data: {
-    classBasePrefix: prefix,
+export interface DropdownMenuProps extends TdDropdownMenuProps {}
+
+@wxComponent()
+export default class DropdownMenu extends SuperComponent {
+  properties = props; // todo: zindex activeColor
+
+  data = {
+    prefix,
     classPrefix: name,
     nodes: null,
     menus: null,
     activeIdx: -1,
     bottom: 0,
-  },
-  relations: {
+  };
+
+  relations: RelationsOptions = {
     './dropdown-item': {
       type: 'child',
     },
-  },
-  methods: {
-    _getAllItems() {
+  };
+
+  lifetimes = {
+    ready() {
+      this.getAllItems();
+    },
+  };
+
+  methods = {
+    getAllItems() {
       const nodes = this.getRelationNodes('./dropdown-item');
-      const menus = nodes.map((a) => {
-        const { title, disabled } = a.data;
-        return { title, disabled };
-      });
+      const menus = nodes.map((a) => a.data);
+
       this.setData({
         nodes,
         menus,
       });
     },
-    _toggleDropdown(e) {
-      const idx = e.target.dataset.index;
-      const { activeIdx } = this.data;
+    toggleDropdown(e: WechatMiniprogram.BaseEvent) {
+      const { index: idx } = e.target.dataset;
+      const { activeIdx, duration } = this.data;
+      const prevItem = this.data.nodes[activeIdx];
+      const currItem = this.data.nodes[idx];
+
+      if (currItem.data.disabled) return;
+
       if (activeIdx !== -1) {
-        this.triggerEvent('close');
-        this.data.nodes[activeIdx].setData({
-          show: false,
-        });
-        this.triggerEvent('closed');
+        prevItem.triggerEvent('close');
+        prevItem.setData(
+          {
+            show: false,
+          },
+          () => {
+            setTimeout(() => {
+              prevItem.triggerEvent('closed');
+            }, duration);
+          },
+        );
       }
+
       if (activeIdx === idx) {
         this.setData({
           activeIdx: -1,
         });
       } else {
-        this.triggerEvent('open');
+        currItem.triggerEvent('open');
         this.setData({
           activeIdx: idx,
         });
-        this.data.nodes[idx].setData({
-          show: true,
-        });
-        this.triggerEvent('opened');
+        currItem.setData(
+          {
+            show: true,
+          },
+          () => {
+            setTimeout(() => {
+              currItem.triggerEvent('opened');
+            }, duration);
+          },
+        );
       }
     },
-  },
-  ready() {
-    this._getAllItems();
-  },
-});
+  };
+}